git-svn-id: https://profuse.googlecode.com/svn/branches/v2@153 aa027e90-d47c-11dd-86d7-074df07e0730

+++ pascal/Date.h
@@ -2,6 +2,7 @@
 #define __PASCAL_DATE_H__
 
 #include <ctime>
+#include <stdint.h>
 
 namespace Pascal {
 
@@ -15,7 +16,7 @@ public:
     Date(unsigned);
     
     operator std::time_t() const;
-    operator unsigned() const;
+    operator uint16_t() const;
     
     unsigned month() const { return _month; }
     unsigned day() const { return _day; }
